rename snapstore to snapmeta
