+++ tests/robustness/snapmeta/snapmeta.go
@@ -1,12 +1,12 @@
-// Package snapstore describes entities that can accept
+// Package snapmeta describes entities that can accept
 // arbitrary metadata and flush it to a persistent repository.
-package snapstore
+package snapmeta
 
 import "github.com/kopia/kopia/tests/robustness/snapif"
 
-// Storer describes the ability to store and retrieve
+// Store describes the ability to store and retrieve
 // a buffer of metadata, indexed by a string key.
-type Storer interface {
+type Store interface {
 	Store(key string, val []byte) error
 	Load(key string) ([]byte, error)
 	Delete(key string)
